Correct tests for reverting to finished builds if no review comment found
Replacing test data with readable set in 'tests/live'. The content should be
equivalent but the python3 bytestream with '\n' as character is replaced by
a readable document. This way the comments are actually parsed differently
which cause a broken test with false assumptions to fail. This is corrected
by adding two test cases for both finding a correct build as well as
reverting to using the last two finished ones if no review comment can be
found with a corresponding new test data set.