Handle empty build / not enough builds
There is an empty build in test data which should be skipped in
get_build_urls_to_compare.

Also handle not enough build error in more cases.

There is a custom exception which should also be used in
the other case where an assert was used in before to
catch and handle the case when, e.g. one product does not
have enough finished builds, which is a legitimate
situation. It will output a TODO-message though as there
is no better plan than the skip the whole product report.
Idea how to handle this better could be to properly
generate a report but report that no changes can be
processed or just do an initial assessment without
comparing builds.