Add new mode for parsing issue references to triage issues
Add options '--bugrefs' as alternative to '-T' plus '--no-empty-sections'.
Based on issue references in openQA, see
https://progress.opensuse.org/projects/openqav3/wiki/Wiki#Test-reviewing,
we can now distinguish between product and openQA issues and keep the rest
without issue references in a 'TODO' section. Additionally, as this can
generate "finished" results, i.e. no further actions by the user should be
necessary, there is an option to not show empty sections which might become
the new default later on.

For the bugreferences readable markdown link markup is used.