Differentiate existing softfails from fails
Previously new softfails have been explicitly mentioned.
Now this is also true for existing softfails which
have been regarded just as "existing issues".

As softfails are assumed to always refer to known product issues we can
automatically move them to the products issues now.

Add option to '--include-softfails'. Default is
changed to not include softfails anymore in the report
as they are less important for the review so it
makes sense to make this configurable.