Fix 'KeyError TODO' with builds having only bugref'd jobs
Sometimes builds are parsed which have no passed or soft-failing jobs and only
failures with bug references so the 'TODO' list would be empty. As it actually
sounds weird to include 'PASSED' jobs into the 'TODO' section this is changed
now to resemble a cleaner 'results_by_bugrefs_dict' which also helps to
reproduce this error locally.