Treat incomplete->failed as new failures
Fixes "missing tests" which can be misleading if just looking at current
builds and the failures don't appear in the report, e.g. because in the
previous build the corresponding job was cancelled or incomplete.

As many more jobs are now considered as "new issues" I am adapting test data
to get rid of some failures to keep at least one test with 'Amber' result and
have 100% statement and branch coverage.