Merge 'dashboard' and 'openqa_review_osd_daily_email'
Extracting common functionality from both scripts we can preserve the
functionality of 'dashboard' but also make use of it in the review report
generated HTML document. Now, 'openqa_review_osd_daily_email' consists of
both the complete report as well as the two filtered reports for
'closed' and 'unassigned' bugs.