Recommend safer installation proposals
The information provided in the README is kind of dangerous
for the unexperienced Linux users, since running any command using
sudo can cause problems which might do not know how to resolve.
Apart from that, using pip and sudo is considered as a very
bad practice that probably causes problems in other projects, thus
most people are using this approach only in a isolated machine
(e.g. VM, Docker, etc) and not into their workstation.

Instead of sudo, a better approach would be to use pip but as a
'normal user', into an isolated environment which is risk-free of
package conflicts. For example, if a user has installed the same
dependency via both zypper and pip, then it will probably end up
with strange problems. This is why VirtualEnv solution is
mentioned in this commit.

Furthermore, it is considered best practice to use the default
way of installing packages, which is the package manager of the
distribution (e.g. zypper) -- given that such packages are available.