Also parse soft-fail needles
Tested with extension of tests as well as against live data for a complete
report. Workaround needles without a proper bugref are ignored gracefully
but with a log message. The corresponding entry is moved to the 'todo'
section as it is not clear if that is really a valid failure and of which
kind.