Fix reminder comment when issue type is not valid
When openQA issue does not have a valid bugref,
it is added as 'product' issue. Reminder comment
script goes through the 'product' issues and looks
for a last comment. It will fail on  assertion if issue
type is not 'bugzilla'.

The commit resolves the issue by skipping
'product' issues with invalid issue type.