Fix workaround bugref not detected on string after bugref
In examples like
https://openqa.opensuse.org/tests/705074#step/start_wayland_plasma5/21
a needle with the name
"generic-desktop-kde-krypton-wayland-boo1059489-pager_not_shown-20180628"
is used which we should consider valid. So far we had been strictly
looking for a number after a bugref separated by "-" which seems to be
too strict. With our bugref only matching for specific prefixes we
should be fine to not need this additional regex part at all.