Add non-ascii characters to the existing tests
The commit adds non-ascii characters
to the data of the existing tests
in order to cover the cases when such
characters are used in subject of
tickets.

openqa-review script should not fail
when the characters are used. All the
characters should be encoded/decoded
properly without fails.

Related issue: https://github.com/okurz/openqa_review/issues/85