Fix sporadic ASCII encoding errors in python2 on complete report
It is hard to fix these errors on python2 and we thought we caught them all
already in before. Hopefully this helps again. I am not really sure because I
could only reproduce the problem based on the complete report generation from
https://openqa.suse.de not relying on any locally saved files. Probably the
content is still encoded differently when read remotely or from files. Each
run took to check takes 40-50 minutes so it was a lengthly fixing process :)