Another try to fix ASCII encoding errors in python2
This time
https://pythonhosted.org/kitchen/unicode-frustrations.html#frustration-3-inconsistent-treatment-of-output
helped me more to understand the real problem :)

Consolidating the codec string to 'utf8' instead of 'utf-8' in all
places.

Also read saved files in utf8 and add unicode test symbols in many local
test files and check rendering into the report files.

Original problem was a symbol "：" but I did not know where to find.
With the previous approach to explicitly encode to utf8 before print I
could create a report but that wraps the whole report with `b'…'`.
However in this way I could find
https://progress.opensuse.org/issues/48077 to be the problematic issue.