Fix "ValueError: Expecting ',' delimiter: line 1..." when trying to parse soft-fail details
This problem happened since about a month during every run but only on
openqa.suse.de data as it seems. The problem was caused by a lot of
escaped text in test module details in conjunction with our code trying
to parse the json document as text first and then parsing as JSON when
escaped characters are already evaluated by python when they should not
have been touched at all. Better we use the `get_json` method of our
browser objects directly to prevent this.