Adapt parsing of soft-failure details on openQA >= 4.5.1528009330.e68ebe2b-3956-gfac9d6dc2
Newer openQA versions have removed again the internal route
"module_components_ajax" so we can not rely on that anymore. Instead we
can query the job details over the corresponding API route and "fake" a
module URL in the format that is expected by the caller of
"_get_url_to_softfailed_module". This is certainly a "lazy" approach to
adapt the code to the behaviour of recent openQA with the least amount
of necessary code change in openqa-review.

Related progress issue: https://progress.opensuse.org/issues/72292