Create PEP440 version string on dev commits to fix problems running tox
As of version 20.3 pip requires PEP440 compliant version strings. Not
generating such a version leads to errors when trying to invoke `tox` in
environments which use a newer pip version.

I've been testing this change on a Git commit which is on a tag and one
that is not on a tag. If it is on a tag the tag is used as version (as
before). If it is not on a tag the most recent tag is used and the
development revision appended, e.g. "1.25.0.dev76" (before it would have
been "1.25.0-76-g777610a").

See
* https://github.com/pypa/pip/issues/9188
* https://progress.opensuse.org/issues/102584