Harmonize they way we query for redmine+bugzilla comments
Previously redmine would read out comments from the issue journal that
was retrieved along with the initial request. For bugzilla the comments
were requested "on-the-fly" when trying to get the "last comment" from
the issue object. To harmonize that and simplify the code I moved the
bugzilla comment reading into the initialization function so that the
actual requests are conducted more near to each other.

This commit changes the test data accordingly as we now need a valid bug
cache object for bug "815". Previously there was already a seemingly
unused cache file for bug "0815", mind the leading zero.